==> 0_static_data_dragon_keys.py <==
#%%
import pandas as pd
from riotwatcher import LolWatcher
import pickle

api_key = "RGAPI-e547be6d-86c3-40f7-a53a-435c3263119c"
watcher = LolWatcher(api_key)
my_region = "na1"


def replace_pickle(path, newFile):
    with open(path, "wb") as f:
        pickle.dump(newFile, f)


#%%
def pull_champion_dict():
    latest = watcher.data_dragon.versions_for_region(my_region)["n"]["champion"]
    static_champ_list = watcher.data_dragon.champions(latest, False, "en_US")

    champ_dict = {}
    for key in static_champ_list["data"]:
        row = static_champ_list["data"][key]
        champ_dict[row["key"]] = row["id"]

    replace_pickle("data/interim/champion_dict.pkl", champ_dict)


def pull_item_dict():
    latest = watcher.data_dragon.versions_for_region(my_region)["n"]["item"]
    static_champ_list = watcher.data_dragon.items(latest, "en_US")

    nameList = []
    goldList = []
    idList = []
    for key in static_champ_list["data"]:
        row = static_champ_list["data"][key]
        idList.append(key)
        nameList.append(row["name"])
        goldList.append(row["gold"]["total"])

    itemDF = pd.DataFrame()
    itemDF["Name"] = nameList
    itemDF["ID"] = idList
    itemDF["Gold"] = goldList

    itemDF.to_csv("data/interim/item_df.csv", index=False)


# %%
pull_champion_dict()
pull_item_dict()
# %%
